Tugeto/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
Django==5.1.6
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-filter==24.3
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
drf-yasg==1.21.10
inflection==0.5.1
itypes==1.2.0
Markdown==3.7
Pillow==11.0.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
python-decouple==3.8
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.8.2
